Move ReST rendering for CLI help into tmt.utils submodule
